refactor(frontend): co-locate tests next to source
Move all tests from frontend/__tests__/ (flat) to siblings of the
source they cover, matching modern React/Vite convention. Drop the
'env-row-addon' / 'stacks-env-roundtrip' filenames in favor of the
source-mirror naming (env-row.test.tsx, form-schema.test.ts) — variant
coverage stays in describe() blocks.

Fixtures move to src/lib/__fixtures__/ alongside the docker-compose
tests that consume them. Imports in moved tests switch from
'../src/...' to relative './...' for same-feature paths and '@/...'
for cross-feature types, matching the codebase's existing import style.

Exclude **/*.test.* and **/__fixtures__/** from tsconfig.app.json so
strict-mode tsc -b doesn't flare on test code that vitest runs at
runtime. No source changes; 124/124 tests still pass.
