+++ frontend/src/lib/__fixtures__/complex-docker-compose.yml
