+++ frontend/src/lib/__fixtures__/invalid-docker-compose.yml
