+++ frontend/src/lib/__fixtures__/simple-docker-compose.yml
